New item category: Reagents, which include spider silk and mithril nugget. Can include also any other items with no apparent use that cannot be eaten. Gems include now pearl and black pearl, comestibles include mandrake root and ginseng root.
Priest now starts with a holy symbol, which is a tool that can be used to turn undead. Turning causes damage to undead and makes them flee, if turning is successful. Also other character classes can try to turn, but it will be more difficult than with priests.

There is now also a reagents shop type that can include gems and other reagents.